[Refactor] Move dictionary cache sink out of Exec
