+++ be/CMakeLists.txt
@@ -583,6 +583,7 @@ set(STARROCKS_LINK_LIBS
     DataSinkResult
     DataSinkExchange
     DataSinkExternal
+    DataSinkDictionaryCache
     DataSinkTablet
     DataSinkFile
     Storage
